# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ReelVault API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VideoInfoRequest {
  /** The video URL to fetch info for */
  url: string;
}

export interface VideoQuality {
  formatId: string;
  quality: string;
  resolution: string;
  ext: string;
  filesize?: number | null;
  isAudioOnly: boolean;
}

export interface VideoInfoResponse {
  title: string;
  thumbnail?: string | null;
  duration?: number | null;
  uploader?: string | null;
  platform: string;
  qualities: VideoQuality[];
  originalUrl: string;
}

export interface DownloadRequest {
  /** The video URL */
  url: string;
  /** The format ID from the qualities list */
  formatId: string;
  /** Whether the user has premium access */
  isPremium?: boolean;
}

export interface DownloadResponse {
  downloadUrl: string;
  filename: string;
  quality: string;
  isAudioOnly: boolean;
}

export interface ErrorResponse {
  error: string;
  message: string;
}