# lib/api-zod/src/generated/types/downloadRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ReelVault API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadRequest {
  /** The video URL */
  url: string;
  /** The format ID from the qualities list */
  formatId: string;
  /** Whether the user has premium access */
  isPremium?: boolean;
}