# lib/api-zod/src/generated/types/downloadResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ReelVault API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadResponse {
  downloadUrl: string;
  filename: string;
  quality: string;
  isAudioOnly: boolean;
}