# lib/api-zod/src/generated/types/videoInfoResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ReelVault API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VideoQuality } from "./videoQuality";

export interface VideoInfoResponse {
  title: string;
  thumbnail?: string | null;
  duration?: number | null;
  uploader?: string | null;
  platform: string;
  qualities: VideoQuality[];
  originalUrl: string;
}