# lib/api-zod/src/generated/types/videoQuality.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ReelVault API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VideoQuality {
  formatId: string;
  quality: string;
  resolution: string;
  ext: string;
  filesize?: number | null;
  isAudioOnly: boolean;
}